Implement VST3 PoC plugin that passes full test validation
This commit proposes a Proof of Concept implementation of a Gain plugin
using the VST3 protocol. Some changes inside the COM dependency were
needed to comply to the protocol. First of all, new COM object instances
should start with 1 reference count instead of 0. Then, there is a need
to have certain interface methods able to take mutable reference to
self.

Some changes were also needed in the specific VST3 interfaces. In
particular, removing from function signatures all (only the ones needed
for now) fat pointers of type `*mut dyn Interface` in favor of raw
pointers of type `*mut c_void`. This is required to be able to create
proper ComPtr objects for calling the COM interface methods needed.

All the following tests were being made on the Windows platform only.
Some minor changes may be needed to achieve the same level of success on
other platforms.

This Proof of Concept implementation is able to pass successfully all 45
tests of the SDK validator. Also, it's able to pass successfully the
Bitwig CLI tests.

As for DAWs support, in order of stability:

- Bitwig 3: No crashes. All features work as expected. The plugin is
able to elaborate sound and control the gain value through the UI.

- Reaper 6: Crashes in some circumstances when deleting the last plugin
instance and then loading a new one. Otherwise works as expected.

- FL Studio 20: Crashes in some circumstances when deleting the plugin
instance. Otherwise works as expected.

- Ableton Live 10: Crashes when touching some parts of the UI when audio
is playing.