Emit cfg() attributes from the proc macro for GUID
Instead of doing this within the macro itself. This fixes the IIDs when
cross compiling, because otherwise it would use the byte order of the
current host system as proc macros are executed by the compiler itself.

Now we just generate both versions and have the compiler pick the right
version when evaluating the generated code.